fixed file structure and browser factory
