+++ server/gradle.properties
@@ -1,3 +1,4 @@
 micronautVersion=2.0.0
 micronautDataVersion=1.0.2
 micronautSecurityVersion=2.0.0
+seleniumVersion=3.141.59
