+++ server/src/test/java/com/objectcomputing/checkins/testcases/TC01_Verify_Upload_File_Success_Msg.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package com.checkInApp.Testcases;
+package com.objectcomputing.checkins.testcases;
 
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.PageFactory;
@@ -9,8 +9,8 @@ import org.testng.annotations.AfterMethod;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
-import com.checkInApp.Pages.HomePage;
-import com.checkInApp.Utility.BrowserFactory;
+import com.objectcomputing.checkins.endtoend.pages.HomePage;
+import com.objectcomputing.checkins.endtoend.utility.BrowserFactory;
 
 /**
  * @author tajr
